requirements.txt
playwright==1.40.0
pandas==2.1.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
streamlit==1.29.0
mechanicalsoup==1.3.0
lxml==4.9.3
requests==2.31.0 